proxylib: Golang L7 filter infrastructure
proxylib is a new top-level directory for Go implemtation of L7 access
control parsers that are compiled into a shared library. This shared
library is loaded at runtime by the Cilium network layer Envoy filter.

The shared library is managed as an Envoy Singleton so that it is only
opened once, by the first redirect that configures it. For this to
work all redirects must pass the same library name and parameters as
these parameters will be only used from the first redirect.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>